History preserving done, moving files around a little bit.
Everything before this commit is history of my private developing
of this project, I preserved it as much as possible, only changed
the commit messages a bit so that they're comprehensible by other
people, and added license notices.

What I'm gonna do next is to import some files that rs depends on,
read more carefully about how to publish on CPAN, follow the conven-
tions, there's much to be done but I'm moving.